pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "chess-lm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.3",
    "python-chess>=1.999",
    "torch>=2.8.0",
    "tqdm>=4.67.1",
    "transformers>=4.56.2",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
include = ["chess_lm*"]

[dependency-groups]
dev = [
    "tiktoken>=0.11.0",
]

[project.scripts]
chess-tokenizer = "chess_lm.tokenizer.cli:main"

[tool.ruff]
exclude = [
    "scripts/",  # Scripts don't need linting
]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]  # Line too long - handled by formatter

[tool.black]
line-length = 120
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 120

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"